add opensearch service validate and tests
