+++ troposphere/opensearchservice.py
@@ -11,6 +11,17 @@ from troposphere import Tags
 
 from . import AWSObject, AWSProperty
 from .validators import boolean, integer
+import re
+
+
+def validate_search_service_engine_version(engine_version):
+    """Validate Engine Version for OpenSearchServiceDomain. The value must be in the format OpenSearch_X.Y or Elasticsearch_X.Y"""
+    engine_version_check = re.compile(r"^(OpenSearch_|Elasticsearch_)\d{1,5}.\d{1,5}")
+    if engine_version_check.match(engine_version) is None:
+        raise ValueError(
+            "OpenSearch EngineVersion must be in the format OpenSearch_X.Y or Elasticsearch_X.Y"
+        )
+    return engine_version
 
 
 class MasterUserOptions(AWSProperty):
@@ -117,7 +128,7 @@ class Domain(AWSObject):
         "DomainName": (str, False),
         "EBSOptions": (EBSOptions, False),
         "EncryptionAtRestOptions": (EncryptionAtRestOptions, False),
-        "EngineVersion": (str, False),
+        "EngineVersion": (validate_search_service_engine_version, False),
         "LogPublishingOptions": (dict, False),
         "NodeToNodeEncryptionOptions": (NodeToNodeEncryptionOptions, False),
         "SnapshotOptions": (SnapshotOptions, False),
